+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Release notes for the Router Craft CMS plugin.
 ### Added
 - Added Craft 3 compatibility.
 - Added shorthand for Year filter's `field` config
+- Added new Date filter with support for year/month/day
 
 
 ## 0.2 - 2015-07-30